+++ src/main/kotlin/dev/voqal/assistant/context/code/ICode.kt
@@ -1,5 +1,6 @@
 package dev.voqal.assistant.context.code
 
+import com.intellij.codeInsight.daemon.impl.HighlightInfo
 import io.vertx.core.json.JsonObject
 
 interface ICode {
@@ -8,4 +9,12 @@ interface ICode {
     val filename: String?
 
     fun toJson(): JsonObject
+
+    fun HighlightInfo.toJson(): JsonObject {
+        return JsonObject().apply {
+            put("description", description)
+            put("severity", severity.toString())
+            //put("line", highlighter.range)
+        }
+    }
 }