+++ src/test/resources/kt/TestFileAccess.kt
@@ -1,0 +1,8 @@
+class TestFileAccess {
+
+    private val a = 2  // multiplier
+    private val b = 5  // offset
+
+    fun calculate(x: Int): Int {
+    }
+}